pyproject.toml
[project]
name = "morning-report"
version = "0.1.0"
description = "Automated daily morning briefing system"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.15",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "requests>=2.31",
]

[project.optional-dependencies]
markets = [
    "yfinance>=0.2",
]
news = [
    "feedparser>=6.0",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.9",
]

[project.scripts]
morning-report = "morning_report.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/morning_report"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
